fix warning

+++ application/CPS/include/BSM_codec.h
@@ -18,7 +18,7 @@ void dump_mem(void *data, int len);
 void bsm_print(BasicSafetyMessage *bsm);
 int bsm_encode(uint8_t **tx_buf, int *tx_buf_len, Obstacle *obstacle);
 int bsm_encode_reg(uint8_t **tx_buf,
-                   int *tx_buf_len,
+                   size_t *tx_buf_len,
                    ObstacleList *obstaclelist);
 void bsm_decode(uint8_t *rx_buf, int rx_buf_len);
